requirements.txt
Flask==2.3.2
moviepy==1.0.3
PySastrawi==1.2.0